ldap: refactor service into separate classes
