+++ db/seeds.rb
@@ -1,4 +1,4 @@
 # Populate the Employee table with current employees
 # Note: this requires being on the campus network. Either onsite or via VPN as
 # it creates an LDAP connection
-Ldap.employees
+Ldap::Queries.employees
